refactor(messages): 拆分消息时间线渲染层并瘦身主组件
将 Messages.tsx 中的 row 子组件、纯渲染工具函数与 timeline 渲染装配层拆分到独立模块，收敛主组件职责。

本次提交包含：

- 新增 MessagesRows.tsx，承接 WorkingIndicator、MessageRow、ReasoningRow、ReviewRow、DiffRow、ExploreRow 等 row 级组件

- 新增 messagesRenderUtils.ts，承接格式化、引擎标记、collapsed 计数、scroll key、debug 开关等纯 helper

- 新增 MessagesTimeline.tsx，承接 renderSingleItem/renderEntry、history sticky header、collapsed indicator、working indicator 等时间线渲染装配

- 收缩 Messages.tsx，仅保留状态编排、滚动与 sticky 计算、timeline 输入组装与外层 shell

验证：

- npx eslint src/features/messages/components/Messages.tsx src/features/messages/components/MessagesTimeline.tsx src/features/messages/components/MessagesRows.tsx src/features/messages/components/messagesRenderUtils.ts

- npm run typecheck

- npx vitest run src/features/messages/components/Messages.test.tsx src/features/messages/components/Messages.live-behavior.test.tsx src/features/messages/components/Messages.windows-render-mitigation.test.tsx

- npm run check:large-files
