+++ src/features/messages/components/MessagesRows.tsx
@@ -1,0 +1,786 @@
+import { memo, useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useTranslation } from "react-i18next";
+import Check from "lucide-react/dist/esm/icons/check";
+import ChevronDown from "lucide-react/dist/esm/icons/chevron-down";
+import ChevronUp from "lucide-react/dist/esm/icons/chevron-up";
+import Copy from "lucide-react/dist/esm/icons/copy";
+import Terminal from "lucide-react/dist/esm/icons/terminal";
+import { AgentIcon } from "../../../components/AgentIcon";
+import type { ConversationItem, QueuedMessage } from "../../../types";
+import { DiffBlock } from "../../git/components/DiffBlock";
+import type { StreamActivityPhase } from "../../threads/hooks/useStreamActivityPhase";
+import { ProxyStatusBadge } from "../../../components/ProxyStatusBadge";
+import { languageFromPath } from "../../../utils/syntax";
+import type { PresentationProfile } from "../presentation/presentationProfile";
+import { parseAgentTaskNotification } from "../utils/agentTaskNotification";
+import { CollapsibleUserTextBlock } from "./CollapsibleUserTextBlock";
+import { ImageLightbox, MessageImageGrid, type MessageImage } from "./MessageMediaBlocks";
+import { Markdown } from "./Markdown";
+import { parseMemoryContextSummary } from "./messagesMemoryContext";
+import { parseReasoning } from "./messagesReasoning";
+import { resolveUserMessagePresentation } from "./messagesUserPresentation";
+import { RuntimeReconnectCard } from "./RuntimeReconnectCard";
+import { resolveAssistantRuntimeReconnectHint } from "./runtimeReconnect";
+import {
+  basenameFromPath,
+  formatDurationMs,
+  MessagesEngine,
+  normalizeAgentTaskStatus,
+  normalizeMessageImageSrc,
+  OPENCODE_NON_STREAMING_HINT_DELAY_MS,
+  resolveAgentTaskDisplaySummary,
+  resolveProvenanceEngineLabel,
+  shouldDisplayWorkingActivityLabel,
+} from "./messagesRenderUtils";
+
+type WorkingIndicatorProps = {
+  isThinking: boolean;
+  proxyEnabled?: boolean;
+  proxyUrl?: string | null;
+  processingStartedAt?: number | null;
+  lastDurationMs?: number | null;
+  heartbeatPulse?: number;
+  hasItems: boolean;
+  reasoningLabel?: string | null;
+  activityLabel?: string | null;
+  activeEngine?: MessagesEngine;
+  waitingForFirstChunk?: boolean;
+  presentationProfile?: PresentationProfile | null;
+  streamActivityPhase?: StreamActivityPhase;
+  primaryLabel?: string | null;
+};
+
+type MessageRowProps = {
+  item: Extract<ConversationItem, { kind: "message" }>;
+  workspaceId?: string | null;
+  threadId?: string | null;
+  isStreaming?: boolean;
+  activeEngine?: MessagesEngine;
+  activeCollaborationModeId?: string | null;
+  enableCollaborationBadge?: boolean;
+  presentationProfile?: PresentationProfile | null;
+  showRuntimeReconnectCard?: boolean;
+  onRecoverThreadRuntime?: (
+    workspaceId: string,
+    threadId: string,
+  ) => Promise<string | null | void> | string | null | void;
+  onRecoverThreadRuntimeAndResend?: (
+    workspaceId: string,
+    threadId: string,
+    message: Pick<QueuedMessage, "text" | "images">,
+  ) => Promise<string | null | void> | string | null | void;
+  retryMessage?: Pick<QueuedMessage, "text" | "images"> | null;
+  isCopied: boolean;
+  onCopy: (
+    item: Extract<ConversationItem, { kind: "message" }>,
+    copyText?: string,
+  ) => void;
+  codeBlockCopyUseModifier?: boolean;
+  onOpenFileLink?: (path: string) => void;
+  onOpenFileLinkMenu?: (event: React.MouseEvent, path: string) => void;
+};
+
+type ReasoningRowProps = {
+  item: Extract<ConversationItem, { kind: "reasoning" }>;
+  workspaceId?: string | null;
+  parsed: ReturnType<typeof parseReasoning>;
+  isExpanded: boolean;
+  isLive: boolean;
+  activeEngine?: MessagesEngine;
+  onToggle: (id: string) => void;
+  onOpenFileLink?: (path: string) => void;
+  onOpenFileLinkMenu?: (event: React.MouseEvent, path: string) => void;
+};
+
+type ReviewRowProps = {
+  item: Extract<ConversationItem, { kind: "review" }>;
+  workspaceId?: string | null;
+  onOpenFileLink?: (path: string) => void;
+  onOpenFileLinkMenu?: (event: React.MouseEvent, path: string) => void;
+};
+
+type DiffRowProps = {
+  item: Extract<ConversationItem, { kind: "diff" }>;
+};
+
+type ExploreRowProps = {
+  item: Extract<ConversationItem, { kind: "explore" }>;
+  isExpanded: boolean;
+  onToggle: (id: string) => void;
+};
+
+function areMessageImagesEqual(
+  previous: Extract<ConversationItem, { kind: "message" }>["images"],
+  next: Extract<ConversationItem, { kind: "message" }>["images"],
+) {
+  if (previous === next) {
+    return true;
+  }
+  if (!previous?.length && !next?.length) {
+    return true;
+  }
+  if (!previous || !next || previous.length !== next.length) {
+    return false;
+  }
+  return previous.every((image, index) => image === next[index]);
+}
+
+function areMessageItemsEqual(
+  previous: Extract<ConversationItem, { kind: "message" }>,
+  next: Extract<ConversationItem, { kind: "message" }>,
+) {
+  return (
+    previous === next ||
+    (
+      previous.id === next.id &&
+      previous.role === next.role &&
+      previous.text === next.text &&
+      previous.engineSource === next.engineSource &&
+      previous.isFinal === next.isFinal &&
+      previous.finalCompletedAt === next.finalCompletedAt &&
+      previous.finalDurationMs === next.finalDurationMs &&
+      previous.selectedAgentName === next.selectedAgentName &&
+      previous.selectedAgentIcon === next.selectedAgentIcon &&
+      areMessageImagesEqual(previous.images, next.images)
+    )
+  );
+}
+
+function areMessageRowPropsEqual(
+  previous: MessageRowProps,
+  next: MessageRowProps,
+) {
+  return (
+    areMessageItemsEqual(previous.item, next.item) &&
+    previous.workspaceId === next.workspaceId &&
+    previous.threadId === next.threadId &&
+    previous.isStreaming === next.isStreaming &&
+    previous.activeEngine === next.activeEngine &&
+    previous.enableCollaborationBadge === next.enableCollaborationBadge &&
+    previous.presentationProfile === next.presentationProfile &&
+    previous.showRuntimeReconnectCard === next.showRuntimeReconnectCard &&
+    previous.onRecoverThreadRuntime === next.onRecoverThreadRuntime &&
+    previous.onRecoverThreadRuntimeAndResend === next.onRecoverThreadRuntimeAndResend &&
+    previous.retryMessage?.text === next.retryMessage?.text &&
+    areMessageImagesEqual(previous.retryMessage?.images, next.retryMessage?.images) &&
+    previous.isCopied === next.isCopied &&
+    previous.onCopy === next.onCopy &&
+    previous.codeBlockCopyUseModifier === next.codeBlockCopyUseModifier &&
+    previous.onOpenFileLink === next.onOpenFileLink &&
+    previous.onOpenFileLinkMenu === next.onOpenFileLinkMenu
+  );
+}
+
+export const WorkingIndicator = memo(function WorkingIndicator({
+  isThinking,
+  proxyEnabled = false,
+  proxyUrl = null,
+  processingStartedAt = null,
+  lastDurationMs = null,
+  heartbeatPulse = 0,
+  hasItems,
+  reasoningLabel = null,
+  activityLabel = null,
+  activeEngine = "claude",
+  waitingForFirstChunk = false,
+  presentationProfile = null,
+  streamActivityPhase = "idle",
+  primaryLabel = null,
+}: WorkingIndicatorProps) {
+  const { t } = useTranslation();
+  const [elapsedMs, setElapsedMs] = useState(0);
+
+  useEffect(() => {
+    if (!isThinking || !processingStartedAt) {
+      setElapsedMs(0);
+      return undefined;
+    }
+    setElapsedMs(Date.now() - processingStartedAt);
+    const interval = window.setInterval(() => {
+      setElapsedMs(Date.now() - processingStartedAt);
+    }, 1000);
+    return () => window.clearInterval(interval);
+  }, [isThinking, processingStartedAt]);
+
+  const showNonStreamingHint =
+    (presentationProfile?.heartbeatWaitingHint ?? activeEngine === "opencode") &&
+    isThinking &&
+    waitingForFirstChunk &&
+    elapsedMs >= OPENCODE_NON_STREAMING_HINT_DELAY_MS;
+  const showActivityLabel = shouldDisplayWorkingActivityLabel(
+    reasoningLabel,
+    activityLabel,
+  );
+  const supportsStreamActivityPhaseFx =
+    activeEngine === "codex" || activeEngine === "claude" || activeEngine === "gemini";
+  const streamPhaseClass =
+    supportsStreamActivityPhaseFx && streamActivityPhase !== "idle"
+      ? ` is-${streamActivityPhase}`
+      : "";
+  const nonStreamingHintText = t("messages.nonStreamingHint");
+  const resolvedNonStreamingHint =
+    nonStreamingHintText === "messages.nonStreamingHint"
+      ? "This model may return non-streaming output, or the network may be unreachable. Please wait..."
+      : nonStreamingHintText;
+  const heartbeatHints = useMemo(() => {
+    const keys = [
+      "messages.opencodeHeartbeatHint1",
+      "messages.opencodeHeartbeatHint2",
+      "messages.opencodeHeartbeatHint3",
+      "messages.opencodeHeartbeatHint4",
+      "messages.opencodeHeartbeatHint5",
+    ];
+    const translated = keys
+      .map((key) => t(key))
+      .filter((value, index) => value !== keys[index]);
+    if (translated.length > 0) {
+      return translated;
+    }
+    return [resolvedNonStreamingHint];
+  }, [resolvedNonStreamingHint, t]);
+  const [heartbeatHintText, setHeartbeatHintText] = useState("");
+  const heartbeatStateRef = useRef<{ lastPulse: number; lastIndex: number }>({
+    lastPulse: 0,
+    lastIndex: -1,
+  });
+
+  useEffect(() => {
+    if (!showNonStreamingHint) {
+      heartbeatStateRef.current = { lastPulse: 0, lastIndex: -1 };
+      setHeartbeatHintText("");
+      return;
+    }
+    if (heartbeatPulse <= 0 || heartbeatPulse === heartbeatStateRef.current.lastPulse) {
+      return;
+    }
+    heartbeatStateRef.current.lastPulse = heartbeatPulse;
+    let randomIndex = Math.floor(Math.random() * heartbeatHints.length);
+    if (heartbeatHints.length > 1 && randomIndex === heartbeatStateRef.current.lastIndex) {
+      randomIndex = (randomIndex + 1) % heartbeatHints.length;
+    }
+    heartbeatStateRef.current.lastIndex = randomIndex;
+    const pulseText = t("messages.opencodeHeartbeatPulse", {
+      pulse: heartbeatPulse,
+      hint: heartbeatHints[randomIndex],
+    });
+    setHeartbeatHintText(
+      pulseText === "messages.opencodeHeartbeatPulse"
+        ? `Heartbeat ${heartbeatPulse}: ${heartbeatHints[randomIndex]}`
+        : pulseText,
+    );
+  }, [heartbeatHints, heartbeatPulse, showNonStreamingHint, t]);
+
+  return (
+    <>
+      {isThinking && (
+        <div className={`working${streamPhaseClass}`}>
+          {proxyEnabled && (
+            <ProxyStatusBadge
+              proxyUrl={proxyUrl}
+              label={t("messages.proxyBadge")}
+              variant="prominent"
+              animated
+              className="working-proxy-badge"
+            />
+          )}
+          <span className="working-spinner" aria-hidden />
+          <div className="working-timer">
+            <span className="working-timer-clock">{formatDurationMs(elapsedMs)}</span>
+          </div>
+          <span className="working-text">
+            {primaryLabel || reasoningLabel || t("messages.generatingResponse")}
+          </span>
+          {showActivityLabel && <span className="working-activity">{activityLabel}</span>}
+          {showNonStreamingHint && (
+            <span className="working-hint">
+              {heartbeatHintText || resolvedNonStreamingHint}
+            </span>
+          )}
+        </div>
+      )}
+      {!isThinking && lastDurationMs !== null && hasItems && (
+        <div className="turn-complete" aria-live="polite">
+          <span className="turn-complete-line" aria-hidden />
+          <span className="turn-complete-label">
+            {t("messages.doneIn", { duration: formatDurationMs(lastDurationMs) })}
+          </span>
+          <span className="turn-complete-line" aria-hidden />
+        </div>
+      )}
+    </>
+  );
+});
+
+export const MessageRow = memo(function MessageRow({
+  item,
+  workspaceId = null,
+  threadId = null,
+  isStreaming = false,
+  activeEngine = "claude",
+  enableCollaborationBadge = false,
+  presentationProfile = null,
+  showRuntimeReconnectCard = false,
+  onRecoverThreadRuntime,
+  onRecoverThreadRuntimeAndResend,
+  retryMessage = null,
+  isCopied,
+  onCopy,
+  codeBlockCopyUseModifier,
+  onOpenFileLink,
+  onOpenFileLinkMenu,
+}: MessageRowProps) {
+  const { t } = useTranslation();
+  const [lightboxIndex, setLightboxIndex] = useState<number | null>(null);
+  const [memorySummaryExpanded, setMemorySummaryExpanded] = useState(false);
+  const [isAgentBadgeExpanded, setIsAgentBadgeExpanded] = useState(false);
+  const userMessagePresentation = useMemo(
+    () => (
+      item.role === "user"
+        ? resolveUserMessagePresentation({
+            text: item.text,
+            selectedAgentName: item.selectedAgentName,
+            selectedAgentIcon: item.selectedAgentIcon,
+            enableCollaborationBadge,
+          })
+        : null
+    ),
+    [
+      enableCollaborationBadge,
+      item.role,
+      item.selectedAgentIcon,
+      item.selectedAgentName,
+      item.text,
+    ],
+  );
+  const memorySummary = useMemo(
+    () =>
+      item.role === "assistant"
+        ? parseMemoryContextSummary(item.text)
+        : userMessagePresentation?.memorySummary ?? null,
+    [item.role, item.text, userMessagePresentation?.memorySummary],
+  );
+  const agentTaskNotification = useMemo(
+    () => parseAgentTaskNotification(item.text),
+    [item.text],
+  );
+  const displayText = agentTaskNotification
+    ? agentTaskNotification.resultText
+    : item.role === "user"
+      ? (userMessagePresentation?.displayText ?? item.text)
+      : memorySummary
+        ? ""
+        : item.text;
+  const selectedAgentName = userMessagePresentation?.selectedAgentName ?? null;
+  const selectedAgentIcon = userMessagePresentation?.selectedAgentIcon ?? null;
+  const hasInjectedAgentPromptBlock =
+    userMessagePresentation?.hasInjectedAgentPromptBlock ?? false;
+  const hasExternalAgentBadge =
+    item.role === "user"
+    && !agentTaskNotification
+    && (Boolean(selectedAgentName) || hasInjectedAgentPromptBlock);
+  const agentTaskDisplay = useMemo(() => {
+    if (!agentTaskNotification) {
+      return null;
+    }
+    return {
+      ...resolveAgentTaskDisplaySummary(agentTaskNotification.summary),
+      status: normalizeAgentTaskStatus(agentTaskNotification.status),
+      outputFileName: basenameFromPath(agentTaskNotification.outputFile),
+    };
+  }, [agentTaskNotification]);
+  useEffect(() => {
+    setIsAgentBadgeExpanded(false);
+  }, [item.id, selectedAgentIcon, selectedAgentName]);
+  const handleToggleAgentBadge = useCallback(() => {
+    setIsAgentBadgeExpanded((current) => !current);
+  }, []);
+  const hasText = displayText.trim().length > 0;
+  const hideCopyButton = item.role === "assistant" && Boolean(memorySummary) && !hasText;
+  const useCodexCanvasMarkdown = presentationProfile
+    ? presentationProfile.codexCanvasMarkdown
+    : activeEngine === "codex";
+  const markdownClassName =
+    item.role === "assistant" && useCodexCanvasMarkdown
+      ? "markdown markdown-codex-canvas"
+      : "markdown";
+  const resolvedMarkdownClassName = isStreaming
+    ? `${markdownClassName} markdown-live-streaming`
+    : markdownClassName;
+  const imageItems = useMemo(() => {
+    if (!item.images || item.images.length === 0) {
+      return [];
+    }
+    return item.images
+      .map((image, index) => {
+        const src = normalizeMessageImageSrc(image);
+        if (!src) {
+          return null;
+        }
+        return { src, label: `Image ${index + 1}` };
+      })
+      .filter(Boolean) as MessageImage[];
+  }, [item.images]);
+  const provenanceLabel = resolveProvenanceEngineLabel(item.engineSource);
+  const runtimeReconnectHint = useMemo(
+    () => (
+      item.role === "assistant"
+        ? resolveAssistantRuntimeReconnectHint(item, Boolean(agentTaskNotification))
+        : null
+    ),
+    [agentTaskNotification, item],
+  );
+
+  const bubbleNode = (
+    <div className={`bubble message-bubble${agentTaskNotification ? " message-bubble-agent-task" : ""}`}>
+      {item.role === "assistant" && provenanceLabel ? (
+        <div className="message-provenance-row">
+          <span className="message-provenance-badge">{provenanceLabel}</span>
+        </div>
+      ) : null}
+      {agentTaskNotification && agentTaskDisplay ? (
+        <div className="message-agent-task-card">
+          <div className="message-agent-task-header">
+            <div className="message-agent-task-avatar" aria-hidden>
+              <AgentIcon
+                seed={agentTaskDisplay.title || agentTaskNotification.taskId || item.id}
+                fallback="codicon-hubot"
+                className="message-agent-task-avatar-icon"
+                size={18}
+              />
+            </div>
+            <div className="message-agent-task-heading">
+              <span className="message-agent-task-eyebrow">Agent session</span>
+              <strong className="message-agent-task-title">{agentTaskDisplay.title}</strong>
+              {agentTaskDisplay.subtitle ? (
+                <span className="message-agent-task-subtitle">{agentTaskDisplay.subtitle}</span>
+              ) : null}
+            </div>
+            <span className={`message-agent-task-status is-${agentTaskDisplay.status.tone}`}>
+              {agentTaskDisplay.status.label}
+            </span>
+          </div>
+          <div className="message-agent-task-meta">
+            {agentTaskNotification.taskId ? (
+              <span className="message-agent-task-chip">task {agentTaskNotification.taskId}</span>
+            ) : null}
+            {agentTaskNotification.toolUseId ? (
+              <span className="message-agent-task-chip">tool {agentTaskNotification.toolUseId}</span>
+            ) : null}
+            {agentTaskDisplay.outputFileName ? (
+              <span className="message-agent-task-chip">{agentTaskDisplay.outputFileName}</span>
+            ) : null}
+          </div>
+        </div>
+      ) : null}
+      {imageItems.length > 0 && (
+        <MessageImageGrid
+          images={imageItems}
+          onOpen={setLightboxIndex}
+          hasText={hasText}
+        />
+      )}
+      {memorySummary ? (
+        <div className="memory-context-summary-card">
+          <button
+            type="button"
+            className="memory-context-summary-toggle"
+            onClick={() => setMemorySummaryExpanded((current) => !current)}
+            aria-expanded={memorySummaryExpanded}
+          >
+            <span className="memory-context-summary-title">
+              {t("messages.memoryContextSummary")}
+            </span>
+            <span className="memory-context-summary-count">
+              {t("messages.memoryContextSummaryCount", {
+                count: memorySummary.lines.length,
+              })}
+            </span>
+            {memorySummaryExpanded ? (
+              <ChevronUp size={14} aria-hidden />
+            ) : (
+              <ChevronDown size={14} aria-hidden />
+            )}
+          </button>
+          {memorySummaryExpanded && (
+            <div className="memory-context-summary-content">
+              {memorySummary.lines.map((line, index) => (
+                <p key={`${item.id}-line-${index}`}>{line}</p>
+              ))}
+            </div>
+          )}
+        </div>
+      ) : null}
+      {runtimeReconnectHint && showRuntimeReconnectCard ? (
+        <RuntimeReconnectCard
+          hint={runtimeReconnectHint}
+          workspaceId={workspaceId}
+          threadId={threadId}
+          onRecoverThreadRuntime={onRecoverThreadRuntime}
+          retryMessage={retryMessage}
+          onRecoverThreadRuntimeAndResend={onRecoverThreadRuntimeAndResend}
+        />
+      ) : null}
+      {hasText && (
+        item.role === "user" && !agentTaskNotification ? (
+          <CollapsibleUserTextBlock content={displayText} />
+        ) : runtimeReconnectHint && showRuntimeReconnectCard ? null : (
+          <Markdown
+            value={displayText}
+            className={resolvedMarkdownClassName}
+            workspaceId={workspaceId}
+            codeBlockStyle="message"
+            codeBlockCopyUseModifier={codeBlockCopyUseModifier}
+            streamingThrottleMs={isStreaming ? 0 : 80}
+            onOpenFileLink={onOpenFileLink}
+            onOpenFileLinkMenu={onOpenFileLinkMenu}
+          />
+        )
+      )}
+      {lightboxIndex !== null && imageItems.length > 0 && (
+        <ImageLightbox
+          images={imageItems}
+          activeIndex={lightboxIndex}
+          onClose={() => setLightboxIndex(null)}
+        />
+      )}
+      {!hideCopyButton && (
+        <button
+          type="button"
+          className={`ghost message-copy-button${isCopied ? " is-copied" : ""}`}
+          onClick={() => onCopy(item, displayText || item.text)}
+          aria-label={t("messages.copyMessage")}
+          title={t("messages.copyMessage")}
+        >
+          <span className="message-copy-icon" aria-hidden>
+            <Copy className="message-copy-icon-copy" size={14} />
+            <Check className="message-copy-icon-check" size={14} />
+          </span>
+        </button>
+      )}
+    </div>
+  );
+
+  const agentBadgeNode = hasExternalAgentBadge ? (
+    <div className={`message-user-agent-rail${isAgentBadgeExpanded ? " is-open" : ""}`}>
+      <button
+        type="button"
+        className="message-agent-icon-button"
+        onClick={handleToggleAgentBadge}
+        aria-expanded={isAgentBadgeExpanded}
+        aria-label={selectedAgentName ? `显示智能体标签：${selectedAgentName}` : "显示智能体标记"}
+        title={selectedAgentName ?? undefined}
+      >
+        <AgentIcon
+          icon={selectedAgentIcon}
+          seed={selectedAgentName ?? item.id}
+          fallback="codicon-hubot"
+          className="message-agent-icon-glyph"
+          size={30}
+        />
+      </button>
+      {isAgentBadgeExpanded && selectedAgentName && (
+        <div className="message-agent-reveal is-visible" role="status">
+          <span className="message-agent-tag-text">{selectedAgentName}</span>
+        </div>
+      )}
+    </div>
+  ) : null;
+
+  return (
+    <div className={`message ${item.role}${agentTaskNotification ? " message-agent-task" : ""}`}>
+      {hasExternalAgentBadge ? (
+        <div className="message-user-layout">
+          {agentBadgeNode}
+          {bubbleNode}
+        </div>
+      ) : bubbleNode}
+    </div>
+  );
+}, areMessageRowPropsEqual);
+
+export const ReasoningRow = memo(function ReasoningRow({
+  item,
+  workspaceId = null,
+  parsed,
+  isExpanded,
+  isLive,
+  activeEngine,
+  onToggle,
+  onOpenFileLink,
+  onOpenFileLinkMenu,
+}: ReasoningRowProps) {
+  const { t } = useTranslation();
+  const { bodyText } = parsed;
+  const shouldPreferRawClaudeContent =
+    activeEngine === "claude" &&
+    item.summary.trim().length > 0 &&
+    item.content.trim().length > 0 &&
+    item.summary.trim() === item.content.trim() &&
+    item.content.includes("\n");
+  const thinkingText = shouldPreferRawClaudeContent
+    ? item.content
+    : bodyText || item.content || item.summary || "";
+  const title = activeEngine === "claude"
+    ? t("messages.thinkingLabel")
+    : isLive
+      ? t("messages.thinkingProcess")
+      : t("messages.thinkingLabel");
+  const provenanceLabel = resolveProvenanceEngineLabel(item.engineSource);
+  return (
+    <div className={`thinking-block${isExpanded ? " is-expanded" : ""}${isLive ? " is-live" : ""}`}>
+      <button
+        type="button"
+        className="thinking-header"
+        onClick={() => onToggle(item.id)}
+      >
+        <span className="thinking-header-copy">
+          <span className="codicon codicon-thinking thinking-glyph" aria-hidden />
+          <span className="thinking-title">{title}</span>
+          {provenanceLabel ? (
+            <span className="message-provenance-badge thinking-provenance-badge">
+              {provenanceLabel}
+            </span>
+          ) : null}
+        </span>
+        <span
+          className={`codicon thinking-icon ${isExpanded ? "codicon-chevron-down" : "codicon-chevron-right"}`}
+          aria-hidden
+        />
+      </button>
+      <div
+        className="thinking-content"
+        style={{ display: isExpanded ? "block" : "none" }}
+      >
+        {thinkingText ? (
+          <div className="reasoning-markdown-surface">
+            <Markdown
+              value={thinkingText}
+              className={`markdown reasoning-markdown${isLive ? " markdown-live-streaming" : ""}`}
+              workspaceId={workspaceId}
+              codeBlockStyle="message"
+              streamingThrottleMs={isLive ? 180 : 80}
+              onOpenFileLink={onOpenFileLink}
+              onOpenFileLinkMenu={onOpenFileLinkMenu}
+            />
+          </div>
+        ) : (
+          <span>{t("messages.noThinkingContent")}</span>
+        )}
+      </div>
+    </div>
+  );
+});
+
+export const ReviewRow = memo(function ReviewRow({
+  item,
+  workspaceId = null,
+  onOpenFileLink,
+  onOpenFileLinkMenu,
+}: ReviewRowProps) {
+  const title = item.state === "started" ? "Review started" : "Review completed";
+  return (
+    <div className="item-card review">
+      <div className="review-header">
+        <span className="review-title">{title}</span>
+        <span
+          className={`review-badge ${item.state === "started" ? "active" : "done"}`}
+        >
+          Review
+        </span>
+      </div>
+      {item.text && (
+        <Markdown
+          value={item.text}
+          className="item-text markdown"
+          workspaceId={workspaceId}
+          onOpenFileLink={onOpenFileLink}
+          onOpenFileLinkMenu={onOpenFileLinkMenu}
+        />
+      )}
+    </div>
+  );
+});
+
+export const DiffRow = memo(function DiffRow({ item }: DiffRowProps) {
+  return (
+    <div className="item-card diff">
+      <div className="diff-header">
+        <span className="diff-title">{item.title}</span>
+        {item.status && <span className="item-status">{item.status}</span>}
+      </div>
+      <div className="diff-viewer-output">
+        <DiffBlock diff={item.diff} language={languageFromPath(item.title)} />
+      </div>
+    </div>
+  );
+});
+
+function exploreKindLabel(kind: ExploreRowProps["item"]["entries"][number]["kind"]) {
+  return (kind[0] ?? "").toUpperCase() + kind.slice(1);
+}
+
+export const ExploreRow = memo(function ExploreRow({
+  item,
+  isExpanded,
+  onToggle,
+}: ExploreRowProps) {
+  const { t } = useTranslation();
+  const title = item.title ?? (item.status === "exploring" ? "Exploring" : "Explored");
+  const isCollapsible =
+    item.collapsible ?? (item.status === "explored" && item.entries.length > 1);
+  const listCollapsed = isCollapsible && !isExpanded;
+  const handleToggle = () => {
+    if (!isCollapsible) {
+      return;
+    }
+    onToggle(item.id);
+  };
+  return (
+    <div className={`tool-inline explore-inline${isCollapsible ? " is-collapsible" : ""}`}>
+      <div className="tool-inline-content">
+        <div className="explore-inline-header">
+          {isCollapsible ? (
+            <button
+              type="button"
+              className="explore-inline-header-toggle"
+              onClick={handleToggle}
+              aria-expanded={isExpanded}
+              aria-label={t("messages.toggleDetails")}
+            >
+              <Terminal
+                className={`tool-inline-icon explore-inline-toggle-icon${
+                  isExpanded ? " is-expanded" : ""
+                } ${item.status === "exploring" ? "processing" : "completed"}`}
+                size={14}
+                aria-hidden
+              />
+              <span className="explore-inline-title">{title}</span>
+            </button>
+          ) : (
+            <>
+              <Terminal
+                className={`tool-inline-icon ${
+                  item.status === "exploring" ? "processing" : "completed"
+                }`}
+                size={14}
+                aria-hidden
+              />
+              <span className="explore-inline-title">{title}</span>
+            </>
+          )}
+        </div>
+        <div className={`explore-inline-list${listCollapsed ? " is-collapsed" : ""}`}>
+          {item.entries.map((entry, index) => (
+            <div key={`${entry.kind}-${entry.label}-${index}`} className="explore-inline-item">
+              <span className="explore-inline-kind">{exploreKindLabel(entry.kind)}</span>
+              <span className="explore-inline-label">{entry.label}</span>
+              {entry.detail && entry.detail !== entry.label && (
+                <span className="explore-inline-detail">{entry.detail}</span>
+              )}
+            </div>
+          ))}
+        </div>
+      </div>
+    </div>
+  );
+});
