+++ src/features/messages/components/MessagesTimeline.tsx
@@ -1,0 +1,472 @@
+import { Fragment, type MutableRefObject, type ReactNode, type RefObject } from "react";
+import { useTranslation } from "react-i18next";
+import Bell from "lucide-react/dist/esm/icons/bell";
+import Flag from "lucide-react/dist/esm/icons/flag";
+import type {
+  AccessMode,
+  ConversationItem,
+  QueuedMessage,
+} from "../../../types";
+import type { GroupedEntry } from "../utils/groupToolItems";
+import { parseAgentTaskNotification } from "../utils/agentTaskNotification";
+import type { PresentationProfile } from "../presentation/presentationProfile";
+import {
+  ToolBlockRenderer,
+  ReadToolGroupBlock,
+  EditToolGroupBlock,
+  BashToolGroupBlock,
+  SearchToolGroupBlock,
+} from "./toolBlocks";
+import {
+  DiffRow,
+  ExploreRow,
+  MessageRow,
+  ReasoningRow,
+  ReviewRow,
+  WorkingIndicator,
+} from "./MessagesRows";
+import { parseReasoning } from "./messagesReasoning";
+import {
+  formatCompletedTimeMs,
+  type HistoryStickyCandidate,
+  type MessagesEngine,
+  resolveProvenanceEngineLabel,
+  shouldHideCodexCanvasCommandCard,
+} from "./messagesRenderUtils";
+
+type MessagesTimelineProps = {
+  activeCollaborationModeId: string | null;
+  activeEngine: MessagesEngine;
+  activeHistoryStickyCandidate: HistoryStickyCandidate | null;
+  activeUserInputRequestId: string | number | null;
+  agentTaskNodeByTaskIdRef: MutableRefObject<Map<string, HTMLDivElement>>;
+  agentTaskNodeByToolUseIdRef: MutableRefObject<Map<string, HTMLDivElement>>;
+  approvalNode: ReactNode;
+  assistantFinalBoundarySet: Set<string>;
+  assistantFinalWithVisibleProcessSet: Set<string>;
+  assistantLiveTurnFinalBoundarySuppressedSet: Set<string>;
+  bottomRef: RefObject<HTMLDivElement | null>;
+  claudeDockedReasoningItems: Array<{
+    item: Extract<ConversationItem, { kind: "reasoning" }>;
+    parsed: ReturnType<typeof parseReasoning>;
+  }>;
+  collapseLiveMiddleStepsEnabled: boolean;
+  collapsedMiddleStepCount: number;
+  codeBlockCopyUseModifier: boolean;
+  copiedMessageId: string | null;
+  effectiveItemsCount: number;
+  expandedItems: Set<string>;
+  groupedEntries: GroupedEntry[];
+  handleCopyMessage: (
+    item: Extract<ConversationItem, { kind: "message" }>,
+    copyText?: string,
+  ) => void;
+  handleExitPlanModeExecuteForItem: (
+    itemId: string,
+    mode: Extract<AccessMode, "default" | "full-access">,
+  ) => Promise<void>;
+  heartbeatPulse: number;
+  isThinking: boolean;
+  isWorking: boolean;
+  lastDurationMs: number | null;
+  latestAssistantMessageId: string | null;
+  latestReasoningLabel: string | null;
+  latestLiveStickyUserMessageId: string | null;
+  latestReasoningId: string | null;
+  latestRetryMessage: Pick<QueuedMessage, "text" | "images"> | null;
+  latestRuntimeReconnectItemId: string | null;
+  latestWorkingActivityLabel: string | null;
+  liveAutoExpandedExploreId: string | null;
+  messageNodeByIdRef: MutableRefObject<Map<string, HTMLDivElement>>;
+  onOpenDiffPath?: (path: string) => void;
+  onRecoverThreadRuntime?: (
+    workspaceId: string,
+    threadId: string,
+  ) => Promise<string | null | void> | string | null | void;
+  onRecoverThreadRuntimeAndResend?: (
+    workspaceId: string,
+    threadId: string,
+    message: Pick<QueuedMessage, "text" | "images">,
+  ) => Promise<string | null | void> | string | null | void;
+  onShowAllHistoryItems: () => void;
+  openFileLink?: (path: string) => void;
+  presentationProfile: PresentationProfile | null;
+  primaryWorkingLabel: string | null;
+  processingStartedAt: number | null;
+  proxyEnabled: boolean;
+  proxyUrl: string | null;
+  reasoningMetaById: Map<string, ReturnType<typeof parseReasoning>>;
+  requestAutoScroll: () => void;
+  selectedExitPlanExecutionByItemKey: Record<string, Extract<AccessMode, "default" | "full-access">>;
+  showFileLinkMenu?: (event: React.MouseEvent, path: string) => void;
+  streamActivityPhase: "idle" | "waiting" | "ingress";
+  threadId: string | null;
+  toggleExpanded: (id: string) => void;
+  userInputNode: ReactNode;
+  visibleCollapsedHistoryItemCount: number;
+  waitingForFirstChunk: boolean;
+  workspaceId: string | null | undefined;
+};
+
+export function MessagesTimeline({
+  activeCollaborationModeId,
+  activeEngine,
+  activeHistoryStickyCandidate,
+  activeUserInputRequestId,
+  agentTaskNodeByTaskIdRef,
+  agentTaskNodeByToolUseIdRef,
+  approvalNode,
+  assistantFinalBoundarySet,
+  assistantFinalWithVisibleProcessSet,
+  assistantLiveTurnFinalBoundarySuppressedSet,
+  bottomRef,
+  claudeDockedReasoningItems,
+  collapseLiveMiddleStepsEnabled,
+  collapsedMiddleStepCount,
+  codeBlockCopyUseModifier,
+  copiedMessageId,
+  effectiveItemsCount,
+  expandedItems,
+  groupedEntries,
+  handleCopyMessage,
+  handleExitPlanModeExecuteForItem,
+  heartbeatPulse,
+  isThinking,
+  isWorking,
+  lastDurationMs,
+  latestAssistantMessageId,
+  latestReasoningLabel,
+  latestLiveStickyUserMessageId,
+  latestReasoningId,
+  latestRetryMessage,
+  latestRuntimeReconnectItemId,
+  latestWorkingActivityLabel,
+  liveAutoExpandedExploreId,
+  messageNodeByIdRef,
+  onOpenDiffPath,
+  onRecoverThreadRuntime,
+  onRecoverThreadRuntimeAndResend,
+  onShowAllHistoryItems,
+  openFileLink,
+  presentationProfile,
+  primaryWorkingLabel,
+  processingStartedAt,
+  proxyEnabled,
+  proxyUrl,
+  reasoningMetaById,
+  requestAutoScroll,
+  selectedExitPlanExecutionByItemKey,
+  showFileLinkMenu,
+  streamActivityPhase,
+  threadId,
+  toggleExpanded,
+  userInputNode,
+  visibleCollapsedHistoryItemCount,
+  waitingForFirstChunk,
+  workspaceId,
+}: MessagesTimelineProps) {
+  const { t } = useTranslation();
+
+  const renderSingleItem = (item: ConversationItem) => {
+    if (item.kind === "message") {
+      const itemRenderKey = `message:${item.id}`;
+      const isCopied = copiedMessageId === item.id;
+      const agentTaskNotification = parseAgentTaskNotification(item.text);
+      const shouldRenderFinalBoundary =
+        item.role === "assistant" &&
+        item.isFinal === true &&
+        assistantFinalBoundarySet.has(item.id) &&
+        !assistantLiveTurnFinalBoundarySuppressedSet.has(item.id);
+      const shouldRenderReasoningBoundary =
+        shouldRenderFinalBoundary && assistantFinalWithVisibleProcessSet.has(item.id);
+      const finalMetaParts: string[] = [];
+      if (typeof item.finalCompletedAt === "number" && item.finalCompletedAt > 0) {
+        finalMetaParts.push(formatCompletedTimeMs(item.finalCompletedAt));
+      }
+      const finalMetaText = finalMetaParts.join(" · ");
+      const bindMessageNode = (node: HTMLDivElement | null) => {
+        if (item.role === "user" && node) {
+          messageNodeByIdRef.current.set(item.id, node);
+        } else {
+          messageNodeByIdRef.current.delete(item.id);
+        }
+        if (agentTaskNotification?.taskId && node) {
+          agentTaskNodeByTaskIdRef.current.set(agentTaskNotification.taskId, node);
+        } else if (agentTaskNotification?.taskId) {
+          agentTaskNodeByTaskIdRef.current.delete(agentTaskNotification.taskId);
+        }
+        if (agentTaskNotification?.toolUseId && node) {
+          agentTaskNodeByToolUseIdRef.current.set(agentTaskNotification.toolUseId, node);
+        } else if (agentTaskNotification?.toolUseId) {
+          agentTaskNodeByToolUseIdRef.current.delete(agentTaskNotification.toolUseId);
+        }
+      };
+      return (
+        <Fragment key={itemRenderKey}>
+          {shouldRenderReasoningBoundary && (
+            <div className="messages-turn-boundary messages-reasoning-boundary" role="separator">
+              <span className="messages-turn-boundary-label">
+                <span className="messages-turn-boundary-label-content">
+                  <Bell className="messages-turn-boundary-icon" size={13} aria-hidden />
+                  <span>推理过程</span>
+                </span>
+              </span>
+              {finalMetaText && (
+                <span
+                  className="messages-turn-boundary-meta messages-turn-boundary-meta-placeholder"
+                  aria-hidden="true"
+                >
+                  {finalMetaText}
+                </span>
+              )}
+            </div>
+          )}
+          <div
+            ref={bindMessageNode}
+            className={[
+              item.id === latestLiveStickyUserMessageId
+                ? "messages-live-sticky-user-message"
+                : "",
+            ]
+              .filter(Boolean)
+              .join(" ") || undefined}
+            data-message-anchor-id={item.id}
+            data-agent-task-id={agentTaskNotification?.taskId ?? undefined}
+            data-agent-tool-use-id={agentTaskNotification?.toolUseId ?? undefined}
+          >
+            <MessageRow
+              item={item}
+              workspaceId={workspaceId}
+              threadId={threadId}
+              isStreaming={
+                activeEngine === "claude" &&
+                isThinking &&
+                item.role === "assistant" &&
+                item.id === latestAssistantMessageId
+              }
+              activeEngine={activeEngine}
+              activeCollaborationModeId={activeCollaborationModeId}
+              enableCollaborationBadge={activeEngine === "codex"}
+              presentationProfile={presentationProfile}
+              showRuntimeReconnectCard={item.id === latestRuntimeReconnectItemId}
+              onRecoverThreadRuntime={onRecoverThreadRuntime}
+              onRecoverThreadRuntimeAndResend={onRecoverThreadRuntimeAndResend}
+              retryMessage={
+                item.id === latestRuntimeReconnectItemId
+                  ? latestRetryMessage
+                  : null
+              }
+              isCopied={isCopied}
+              onCopy={handleCopyMessage}
+              codeBlockCopyUseModifier={codeBlockCopyUseModifier}
+              onOpenFileLink={openFileLink}
+              onOpenFileLinkMenu={showFileLinkMenu}
+            />
+          </div>
+          {shouldRenderFinalBoundary && (
+            <div className="messages-turn-boundary messages-final-boundary" role="separator">
+              <span className="messages-turn-boundary-label">
+                <span className="messages-turn-boundary-label-content">
+                  <Flag className="messages-turn-boundary-icon" size={13} aria-hidden />
+                  <span>最终消息</span>
+                </span>
+              </span>
+              {finalMetaText && (
+                <span className="messages-turn-boundary-meta">{finalMetaText}</span>
+              )}
+            </div>
+          )}
+        </Fragment>
+      );
+    }
+    if (item.kind === "reasoning") {
+      const itemRenderKey = `reasoning:${item.id}`;
+      const isExpanded = expandedItems.has(item.id);
+      const parsed = reasoningMetaById.get(item.id) ?? parseReasoning(item);
+      const isLiveReasoning =
+        isThinking && latestReasoningId === item.id;
+      return (
+        <ReasoningRow
+          key={itemRenderKey}
+          item={item}
+          workspaceId={workspaceId}
+          parsed={parsed}
+          isExpanded={isExpanded}
+          isLive={isLiveReasoning}
+          activeEngine={activeEngine}
+          onToggle={toggleExpanded}
+          onOpenFileLink={openFileLink}
+          onOpenFileLinkMenu={showFileLinkMenu}
+        />
+      );
+    }
+    if (item.kind === "review") {
+      return (
+        <ReviewRow
+          key={`review:${item.id}`}
+          item={item}
+          workspaceId={workspaceId}
+          onOpenFileLink={openFileLink}
+          onOpenFileLinkMenu={showFileLinkMenu}
+        />
+      );
+    }
+    if (item.kind === "diff") {
+      return <DiffRow key={`diff:${item.id}`} item={item} />;
+    }
+    if (item.kind === "tool") {
+      if (shouldHideCodexCanvasCommandCard(item, activeEngine)) {
+        return null;
+      }
+      const isExpanded = expandedItems.has(item.id);
+      const selectedExitPlanExecutionMode =
+        selectedExitPlanExecutionByItemKey[`${threadId ?? "no-thread"}:${item.id}`] ?? null;
+      const provenanceLabel = resolveProvenanceEngineLabel(item.engineSource);
+      return (
+        <div key={`tool:${item.id}`} className="message-tool-block-shell">
+          {provenanceLabel ? (
+            <div className="message-provenance-row">
+              <span className="message-provenance-badge">{provenanceLabel}</span>
+            </div>
+          ) : null}
+          <ToolBlockRenderer
+            item={item}
+            workspaceId={workspaceId}
+            isExpanded={isExpanded}
+            onToggle={toggleExpanded}
+            onRequestAutoScroll={requestAutoScroll}
+            activeCollaborationModeId={activeCollaborationModeId}
+            activeEngine={activeEngine}
+            hasPendingUserInputRequest={activeUserInputRequestId !== null}
+            onOpenDiffPath={onOpenDiffPath}
+            selectedExitPlanExecutionMode={selectedExitPlanExecutionMode}
+            onExitPlanModeExecute={handleExitPlanModeExecuteForItem}
+          />
+        </div>
+      );
+    }
+    if (item.kind === "explore") {
+      const isExpanded = liveAutoExpandedExploreId === item.id || expandedItems.has(item.id);
+      return (
+        <ExploreRow
+          key={`explore:${item.id}`}
+          item={item}
+          isExpanded={isExpanded}
+          onToggle={toggleExpanded}
+        />
+      );
+    }
+    return null;
+  };
+
+  const renderEntry = (entry: GroupedEntry) => {
+    if (entry.kind === "readGroup") {
+      const firstItem = entry.items[0];
+      return <ReadToolGroupBlock key={`rg-${firstItem?.id ?? "read-group"}`} items={entry.items} />;
+    }
+    if (entry.kind === "editGroup") {
+      const firstItem = entry.items[0];
+      return (
+        <EditToolGroupBlock
+          key={`eg-${firstItem?.id ?? "edit-group"}`}
+          items={entry.items}
+          onOpenDiffPath={onOpenDiffPath}
+        />
+      );
+    }
+    if (entry.kind === "bashGroup") {
+      if (activeEngine === "codex" || activeEngine === "claude") {
+        return null;
+      }
+      const firstItem = entry.items[0];
+      return (
+        <BashToolGroupBlock
+          key={`bg-${firstItem?.id ?? "bash-group"}`}
+          items={entry.items}
+          onRequestAutoScroll={requestAutoScroll}
+        />
+      );
+    }
+    if (entry.kind === "searchGroup") {
+      const firstItem = entry.items[0];
+      return <SearchToolGroupBlock key={`sg-${firstItem?.id ?? "search-group"}`} items={entry.items} />;
+    }
+    return renderSingleItem(entry.item);
+  };
+
+  return (
+    <>
+      {activeHistoryStickyCandidate && (
+        <div
+          className="messages-history-sticky-header"
+          data-history-sticky-message-id={activeHistoryStickyCandidate.id}
+          aria-hidden="true"
+        >
+          <div className="messages-history-sticky-header-inner">
+            <div className="messages-history-sticky-header-content">
+              <div className="messages-history-sticky-header-bubble">
+                <div className="messages-history-sticky-header-text">
+                  {activeHistoryStickyCandidate.text}
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      )}
+      <div className="messages-full">
+        {visibleCollapsedHistoryItemCount > 0 && (
+          <div
+            className="messages-collapsed-indicator"
+            data-collapsed-count={visibleCollapsedHistoryItemCount}
+            onClick={onShowAllHistoryItems}
+          >
+            {t("messages.showEarlierMessages", { count: visibleCollapsedHistoryItemCount })}
+          </div>
+        )}
+        {groupedEntries.map(renderEntry)}
+        {claudeDockedReasoningItems.map(({ item, parsed }) => (
+          <ReasoningRow
+            key={`claude-live-${item.id}`}
+            item={item}
+            workspaceId={workspaceId}
+            parsed={parsed}
+            isExpanded={isThinking && latestReasoningId === item.id ? true : expandedItems.has(item.id)}
+            isLive={isThinking && latestReasoningId === item.id}
+            onToggle={toggleExpanded}
+            onOpenFileLink={openFileLink}
+            onOpenFileLinkMenu={showFileLinkMenu}
+          />
+        ))}
+        {userInputNode}
+        {isThinking && collapseLiveMiddleStepsEnabled && collapsedMiddleStepCount > 0 && (
+          <div className="messages-live-middle-collapsed-indicator" role="status">
+            {t("messages.middleStepsCollapsedHint", { count: collapsedMiddleStepCount })}
+          </div>
+        )}
+        <WorkingIndicator
+          isThinking={isWorking}
+          proxyEnabled={proxyEnabled}
+          proxyUrl={proxyUrl}
+          processingStartedAt={processingStartedAt}
+          lastDurationMs={lastDurationMs}
+          heartbeatPulse={heartbeatPulse}
+          hasItems={effectiveItemsCount > 0}
+          reasoningLabel={latestReasoningLabel}
+          activityLabel={latestWorkingActivityLabel}
+          primaryLabel={primaryWorkingLabel}
+          activeEngine={activeEngine}
+          waitingForFirstChunk={waitingForFirstChunk}
+          presentationProfile={presentationProfile}
+          streamActivityPhase={streamActivityPhase}
+        />
+        {!effectiveItemsCount && !userInputNode && (
+          <div className="empty messages-empty">
+            {t("messages.emptyThread")}
+          </div>
+        )}
+        {approvalNode}
+        <div ref={bottomRef} />
+      </div>
+    </>
+  );
+}
